Update to allow new server
Update to allow for running the service independently on it's own web server.

If a port is provided the service will attempt to start the service on a new
express server instance.

It checks if the port is available and throws otherwise.

Also remove app as an input to the public methods, app is now encapsulated
in the web service injected at construction time or if a port is provided
at install time.